+++ pyvo/dal/sla.py
@@ -24,8 +24,12 @@ endpoint.
 from __future__ import (
     absolute_import, division, print_function, unicode_literals)
 
+from pyvo.io.vosi.vodataservice import TableParam
+
 from astropy.units import Quantity, Unit
 from astropy.units import spectral as spectral_equivalencies
+from astropy.io.votable.tree import Field
+from astropy.table import Table
 
 from .query import DALResults, DALQuery, DALService, Record
 
@@ -80,6 +84,45 @@ class SLAService(DALService):
            the base URL for submitting search queries to the service.
         """
         super(SLAService, self).__init__(baseurl,)
+
+    def _get_metadata(self):
+        """
+        download the metadata resource
+        """
+        if not hasattr(self, "_metadata"):
+            query = self.create_query(request='getCapabilities')
+            metadata = query.execute_votable()
+
+            setattr(self, "_metadata", metadata)
+
+    @property
+    def description(self):
+        """
+        the service description.
+        """
+        self._get_metadata()
+
+        try:
+            return getattr(self, "_metadata", None).description
+        except AttributeError:
+            return None
+
+    @property
+    def columns(self):
+        """
+        the available columns on this service
+        """
+        self._get_metadata()
+        fields = filter(
+            lambda field_or_param: isinstance(field_or_param, Field),
+            self._metadata.iter_fields_and_params()
+        )
+
+        try:
+            return [
+                TableParam.from_field(field) for field in fields]
+        except AttributeError:
+            return []
 
     def search(self, wavelength, **keywords):
         """
@@ -149,6 +192,34 @@ class SLAService(DALService):
         """
         return SLAQuery(self.baseurl, wavelength, request, **keywords)
 
+    def describe(self):
+        print(self.description)
+        print()
+
+        rows = [(
+            col.name,
+            col.description,
+            col.unit,
+            col.ucd,
+            col.utype,
+            col.datatype.arraysize,
+            col.datatype.content,
+        ) for col in self.columns]
+
+        names = (
+            'name',
+            'description',
+            'unit',
+            'ucd',
+            'utype',
+            'arraysize',
+            'datatype',
+        )
+
+        table = Table(rows=rows, names=names)
+        table.pprint(
+            max_lines=-1, max_width=-1, show_unit=False, show_dtype=False)
+
 
 class SLAQuery(DALQuery):
     """
@@ -190,7 +261,7 @@ class SLAQuery(DALQuery):
         """
         super(SLAQuery, self).__init__(baseurl)
 
-        if wavelength:
+        if wavelength is not None:
             self.wavelength = wavelength
 
         self.request = request