+++ pyvo/dal/tap.py
@@ -12,6 +12,7 @@ from distutils.version import LooseVersion
 
 import requests
 
+from astropy.extern import six
 from astropy.io.votable import parse as votableparse
 
 from .query import (
@@ -22,6 +23,8 @@ from .adhoc import DatalinkResultsMixin, DatalinkRecordMixin, SodaRecordMixin
 
 from ..io import vosi, uws
 from ..io.vosi import tapregext as tr
+
+from ..utils.formatting import para_format_desc
 from ..utils.http import session as s
 
 __all__ = [
@@ -296,6 +299,33 @@ class TAPService(DALService, AvailabilityMixin, CapabilityMixin):
         """
         return TAPQuery(
             self.baseurl, query, mode, language, maxrec, uploads, **keywords)
+
+    def describe(self, width=None):
+        """
+        Print a summary description of this service.
+
+        This includes the interface capabilities, and the content description
+        if it doesn't contains multiple data collections (in other words, it is
+        not a TAP service).
+        """
+        if len(self.tables) == 1:
+            description = next(self.tables.values()).description
+
+            if width:
+                description = para_format_desc(description, width)
+
+            print(description)
+            print()
+
+        capabilities = filter(
+            lambda x: not six.text_type(x.standardid).startswith(
+                'ivo://ivoa.net/std/VOSI'),
+            self.capabilities
+        )
+
+        for cap in capabilities:
+            cap.describe()
+            print()
 
 
 class AsyncTAPJob(object):